Crea los métodos para buscar en la base de datos los libros buscados y los autores de los respectivos libros

+++ src/main/java/com/project/LiterAlura/model/DatosLibro.java
@@ -10,7 +10,7 @@ public record DatosLibro(
         @JsonAlias("title") String titulo,
         @JsonAlias("authors") List<DatosAutor> autor,
         @JsonAlias("languages") List<String> idioma,
-        @JsonAlias("download_count") Double descargas) {
+        @JsonAlias("download_count") Integer descargas) {
 
     @Override
     public String toString() {
